refactor: Finish VintageController.createOrder

+++ src/test/java/com/marketplace/vintage/expression/Exp4jExpressionSolverTest.java
@@ -28,7 +28,7 @@ public class Exp4jExpressionSolverTest {
     @Test
     void defaultExpression() {
         String expression = VintageConstants.DEFAULT_EXPEDITION_PRICE_EXPRESSION_STRING;
-        BigDecimal result = expressionSolver.solve(expression, Map.of("itemsPrice", 100.0, "tax", 0.1));
+        BigDecimal result = expressionSolver.solve(expression, Map.of("itemsPrice", BigDecimal.valueOf(100.0), "tax", BigDecimal.valueOf(0.1)));
         assertEquals(BigDecimal.valueOf(9.9).compareTo(result), 0);
     }
